Request based commit caching in the CommitService
Allows caching of commits throughout the whole request, in the
RequestStore. This is cross instance, and also caches all commits
fetched from other RPCs.

In testing, mostly the Ci::Pipeline had failures, as it referenced a
commit but did not create a repository. Errors were of GRPC::NotFound
suggesting that the commit wasn't found. These are now thrown as
Gitlab::Git::NoRepository errors to signal whats happening.