Use an existing SHA for seeding Ci::Pipeline
Prior to this change, a bogus SHA was used, which wasn't found by
Gitaly or any other way as its not in gitlab-test.

This change makes this right. The commit chosen was merged to master
about 2 months ago in master. Master is the ref for the default
pipelines too, so that finally matches as well.

Found this a couple of times, but worked around it, recently
https://gitlab.com/gitlab-org/gitlab-ce/merge_requests/17553 was the
straw that broke the camels back.

Tests showed this does one query more, to fetch metadata for commits.