Trigger CI with Gitaly self healing hooks
This commit triggers the GitLab CE pipeline so it can be tested if it
works with Gitaly that will self heal the hooks. My understanding is
that it won't, as now the hooks directory is removed when a project is
seeded.

Gitaly MR: https://gitlab.com/gitlab-org/gitaly/merge_requests/886

Part of: https://gitlab.com/gitlab-org/gitaly/issues/1226