Add timeouts for each RPC call
When setting the timeouts the client can give up on the RPC call, which
allows for a protection for the client and server.

This commit also removes the implicit lack of a timeout when running on
sidekiq. This is a good thing as if one server is overloaded with work,
it wouldn't block the whole queue from processing.

Part of: https://gitlab.com/groups/gitlab-org/-/epics/1737